Optimized function `bitShiftBlob` to shift data in 64-bit chunks
The shifting is now done as units of 64-bit words rather than
byte-wise.

This should be faster since we loop fewer times and we leverage the
inherent CPU hardware for propagating carry bits for the majority of
the bytes in the buffer (we still have to do carries at word-edges but
that's better than doing it for every byte!).