Fix for C++ UB in scriptnum_tests introduced by a recent commit
We were overflowing in the top-most part of the test when doing e.g. `a + b`
after adding the ability of the tests to explore more 64-bit values than
before.

Fixed by checking for overflow before doing e.g. `a + b`, and skipping
tests that rely on an overflowing expression.